+++ CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Unreleased]
+
+### Added
+- import of ClazzStatistic.csv can now be triggered by JMX (e.g. with `jconsole`)
+
+
 ## [2.6.5] (22-Jun-2025)
 
 ### Fixed